Move method factories to factories/model-editor
