+++ extensions/ql-vscode/src/stories/model-editor/MethodName.stories.tsx
@@ -3,7 +3,7 @@ import * as React from "react";
 import { Meta, StoryFn } from "@storybook/react";
 
 import { MethodName as MethodNameComponent } from "../../view/model-editor/MethodName";
-import { createMethod } from "../../../test/factories/data-extension/method-factories";
+import { createMethod } from "../../../test/factories/model-editor/method-factories";
 
 export default {
   title: "CodeQL Model Editor/Method Name",
