+++ extensions/ql-vscode/src/view/model-editor/__tests__/ModeledMethodsList.spec.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { render as reactRender, screen } from "@testing-library/react";
-import { createMethod } from "../../../../test/factories/data-extension/method-factories";
+import { createMethod } from "../../../../test/factories/model-editor/method-factories";
 import { InProgressMethods } from "../../../model-editor/shared/in-progress-methods";
 import { createMockExtensionPack } from "../../../../test/factories/model-editor/extension-pack";
 import { Mode } from "../../../model-editor/shared/mode";
