+++ extensions/ql-vscode/test/factories/model-editor/method-factories.ts
