+++ extensions/ql-vscode/test/vscode-tests/no-workspace/model-editor/methods-usage/methods-usage-data-provider.test.ts
@@ -5,7 +5,7 @@ import { DatabaseItem } from "../../../../../src/databases/local-databases";
 import {
   createMethod,
   createUsage,
-} from "../../../../factories/data-extension/method-factories";
+} from "../../../../factories/model-editor/method-factories";
 import { mockedObject } from "../../../utils/mocking.helpers";
 
 describe("MethodsUsageDataProvider", () => {
