+++ extensions/ql-vscode/test/vscode-tests/no-workspace/model-editor/methods-usage/methods-usage-panel.test.ts
@@ -7,7 +7,7 @@ import { mockedObject } from "../../../utils/mocking.helpers";
 import {
   createMethod,
   createUsage,
-} from "../../../../factories/data-extension/method-factories";
+} from "../../../../factories/model-editor/method-factories";
 
 describe("MethodsUsagePanel", () => {
   const mockCliServer = mockedObject<CodeQLCliServer>({});
